Moved CSS overflow:hidden to only apply if dynamic is true
Overflow:hidden is used to prevent the possibility of double scroll bars on the body when a tall, scroll modal (i.e. data-dynamic==true) is used on a tall page. For this, it works great. But for small, non-dynamic modals it's better not to apply overflow:hidden because: 1.) it's not necessary (given we never have to worry about there being two body scroll bars because it's a small sized modal afterall) and 2.) using it for small modals causes the page to 'jump' and shift to the right b/c tall pages go from having a scroll bar to not having a scroll bar. By applying overflow:hidden only when data-dynamic==true, we avoid this jarring page shift for small-sized modals. 